bot/requirements.txt
aiogram-2.17.1
pandas-1.3.4
fastparquet-0.7.2
h3-3.7.3